kvformat: drop KVScannerGeneric
This patch gets rid off KVScannerGeneric in favour of the boosted up
KVScannerSimple.
