+++ modules/kvformat/Makefile.am
@@ -7,10 +7,6 @@ modules_kvformat_libkvformat_la_SOURCES=	\
 	modules/kvformat/kv-parser.h		\
 	modules/kvformat/kv-scanner.c		\
 	modules/kvformat/kv-scanner.h		\
-	modules/kvformat/kv-scanner-generic.c \
-	modules/kvformat/kv-scanner-generic.h \
-	modules/kvformat/kv-scanner-simple.c \
-	modules/kvformat/kv-scanner-simple.h \
 	modules/kvformat/linux-audit-parser.c	\
 	modules/kvformat/linux-audit-parser.h	\
 	modules/kvformat/kv-parser-grammar.y	\
